docs:create constants.js file and transferred the constants into it
