+++ js/constants.js
@@ -1,0 +1,63 @@
+
+const descriptions = [
+  "Солнечный закат и влюбленные в него",
+  "Доброе утро из рая",
+  "Когда жизнь дает цитрины, делайте лимонад",
+  "Городские огни и бесконечные ночи",
+  "Тропическое настроение",
+  "Песочные ноги и загорелые носы",
+  "Жизнь лучше в тапочках",
+  "Минутка спокойствия в суматохе",
+  "Полдень солнца",
+  "Морское приключение",
+  "Утро с кофе, книгами и пуфиком",
+  "Приключения ждут снаружи",
+  "Солнце, море и пачка картошки фри",
+  "Смех - лучший медицинский препарат",
+  "Только хорошие волны",
+  "Утренняя заря",
+  "Теплое свечение вечера",
+  "Солнце в волосах",
+  "Горячие напитки и холодные зимние вечера",
+  "Рассказы из океана",
+  "В поисках приключений",
+  "Солнечные улыбки",
+  "Теплые ночи и теплые огни",
+  "Мечты в горах",
+  "Жизнь - пляж, и еще что-то",
+  "Кусочек неба на земле",
+  "Городские улицы и городской бит",
+  "Душа дикой цветочной",
+  "Сохраняя момент",
+  "Куда заведет тебя дорога",
+  "Свобода в ветре",
+  "Капля цвета в сером мире",
+  "Кофе и страсть к путешествиям",
+  "Ненаписанные истории и нерассказанные сказки",
+];
+const USERS_ARRAY_LENGTH = 25;
+const MAX_LENGTH_MESSAGE = 2;
+const MAX_NUM_PHOTOS = 6;
+const messageForComments = [
+  "Всё отлично!",
+  "В целом всё неплохо. Но не всё.",
+  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
+
+];
+const names = [
+  "Артем",
+  "Виктор",
+  "Петр",
+  "Антонина",
+  "Кирилл",
+  "Валерия",
+  "Настя",
+  "Сеня",
+  "Макс",
+  "Егор",
+  "Изольда",
+];
+export{USERS_ARRAY_LENGTH,MAX_LENGTH_MESSAGE,MAX_NUM_PHOTOS,messageForComments,names,descriptions};
