+++ js/main.js
@@ -1,5 +1,3 @@
-import {randomNum} from "./random-num.js";
-import {createCommentsArray} from "./comment-generator.js";
 import {createPhotoParametrsArray} from "./photo-data-generator.js";
 
 console.log(createPhotoParametrsArray());
