+++ js/photo-data-generator.js
@@ -1,43 +1,6 @@
-import {randomNum} from "./random-num.js";
-import {createCommentsArray} from "./comment-generator.js";
-const descriptions = [
-  "Солнечный закат и влюбленные в него",
-  "Доброе утро из рая",
-  "Когда жизнь дает цитрины, делайте лимонад",
-  "Городские огни и бесконечные ночи",
-  "Тропическое настроение",
-  "Песочные ноги и загорелые носы",
-  "Жизнь лучше в тапочках",
-  "Минутка спокойствия в суматохе",
-  "Полдень солнца",
-  "Морское приключение",
-  "Утро с кофе, книгами и пуфиком",
-  "Приключения ждут снаружи",
-  "Солнце, море и пачка картошки фри",
-  "Смех - лучший медицинский препарат",
-  "Только хорошие волны",
-  "Утренняя заря",
-  "Теплое свечение вечера",
-  "Солнце в волосах",
-  "Горячие напитки и холодные зимние вечера",
-  "Рассказы из океана",
-  "В поисках приключений",
-  "Солнечные улыбки",
-  "Теплые ночи и теплые огни",
-  "Мечты в горах",
-  "Жизнь - пляж, и еще что-то",
-  "Кусочек неба на земле",
-  "Городские улицы и городской бит",
-  "Душа дикой цветочной",
-  "Сохраняя момент",
-  "Куда заведет тебя дорога",
-  "Свобода в ветре",
-  "Капля цвета в сером мире",
-  "Кофе и страсть к путешествиям",
-  "Ненаписанные истории и нерассказанные сказки",
-];
-const USERS_ARRAY_LENGTH = 25;
-
+import{USERS_ARRAY_LENGTH,descriptions} from "./constants.js";
+import{randomNum} from "./utils.js";
+import{createCommentsArray} from "./comment-generator.js";
 function createPhotoParametrsArray() {
   const usersArray = [];
   for (let i = 1; i <= USERS_ARRAY_LENGTH ; i++) {
@@ -51,4 +14,4 @@ function createPhotoParametrsArray() {
   }
   return usersArray;
 }
-export{createPhotoParametrsArray};
+export {createPhotoParametrsArray};
