+++ js/utils.js
@@ -1,4 +1,5 @@
 function randomNum(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 }
+
 export {randomNum};
